refactor: move test into src
